starting to delve into how to call the perl function from c...
the example at <https://perldoc.perl.org/perlcall#Returning-a-Scalar> shows how to give the perl-function a list of arguments for @_, and how to retrieve the scalar it returns.  The next examples below show how to return a whole list rather than just a single scalar.  I _think_ all I need is a single scalar, because that's all that DlgProc itself returns...  Note the various macros used for establishing a local stack, and playing nice with temporary SVs